requirements.txt
Django==5.2.8
djangorestframework==3.15.2
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
